Sum function modified to accept double value

+++ Calculator.java
@@ -1,6 +1,6 @@
 public class Calculator{
-	public void sum(int a, int b){
-		int sum = a + b;
+	public void sum(double a, double b){
+		double sum = a + b;
 		System.out.println(sum);
 	}
 }
